feat: route get activity booking implemented and new files added
