+++ src/app.ts
@@ -17,7 +17,8 @@ import {
   paymentsRouter,
   hotelsRouter,
   bookingRouter,
-  activyRouter
+  activyRouter,
+  activityBookingRouter
 } from "@/routers";
 
 const app = express();
@@ -34,6 +35,7 @@ app
   .use("/hotels", hotelsRouter)
   .use("/booking", bookingRouter)
   .use("/activity", activyRouter)
+  .use("/activitybooking", activityBookingRouter)
   .use(handleApplicationErrors);
 
 export function init(): Promise<Express> {
